# packages/api-contracts/rest/openapi.yaml
openapi: 3.1.0
info:
  title: ManaBandhu REST API
  version: 1.0.0
servers:
  - url: /api/v1
paths:
...
  /safety/reports:
    get:
      operationId: listSafetyReports
      responses:
        '200':
          description: Safety reports
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SafetyReport'
    post:
      operationId: createSafetyReport
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateSafetyReportInput'
      responses:
        '201':
          description: Report created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SafetyReport'
  /safety/reports/{id}/resolve:
    post:
      operationId: resolveSafetyReport
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '204':
          description: Report resolved
  /safety/blocked-users:
    get:
      operationId: listBlockedUsers
      responses:
        '200':
          description: Blocked users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/BlockedUser'
    post:
      operationId: blockUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BlockUserInput'
      responses:
        '201':
          description: User blocked
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BlockedUser'
  /safety/blocked-users/{blockedUserId}/unblock:
    post:
      operationId: unblockUser
      parameters:
        - name: blockedUserId
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '204':
          description: User unblocked
  /safety/trusted-contacts:
    get:
      operationId: listTrustedContacts
      responses:
        '200':
          description: Trusted contacts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TrustedContact'
    post:
      operationId: createTrustedContact
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTrustedContactInput'
      responses:
        '201':
          description: Contact created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TrustedContact'
  /utilities/packages:
    get:
      operationId: listPackageTrackings
      responses:
        '200':
          description: Package trackings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PackageTracking'
    post:
      operationId: createPackageTracking
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePackageTrackingInput'
      responses:
        '201':
          description: Tracking created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PackageTracking'
  /utilities/nearby:
    get:
      operationId: listNearbyPlaces
      responses:
        '200':
          description: Nearby places
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/NearbyPlace'
      parameters:
        - name: category
          in: query
          schema: { type: string }
  /utilities/emergency:
    get:
      operationId: listEmergencyResources
      responses:
        '200':
          description: Emergency resources
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/EmergencyResource'
      parameters:
        - name: category
          in: query
          schema: { type: string }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    SafetyReport:
      type: object
      additionalProperties: false
      required: [id, reporterId, targetId, category, description, status, severity, createdAt]
      properties:
        id: { type: string, format: uuid }
        reporterId: { type: string, format: uuid }
        targetId: { type: string, format: uuid }
        category: { type: string, maxLength: 80 }
        description: { type: string, maxLength: 4000 }
        status: { type: string, enum: [OPEN, IN_REVIEW, RESOLVED, DISMISSED] }
        severity: { type: string, enum: [LOW, MEDIUM, HIGH, CRITICAL] }
        evidenceUrls: { type: string, maxLength: 2000 }
        createdAt: { type: string, format: date-time }
        resolvedAt: { type: string, format: date-time }
    CreateSafetyReportInput:
      type: object
      additionalProperties: false
      required: [targetId, category, description, severity]
      properties:
        targetId: { type: string, format: uuid }
        category: { type: string, maxLength: 80 }
        description: { type: string, maxLength: 4000 }
        severity: { type: string, enum: [LOW, MEDIUM, HIGH, CRITICAL] }
        evidenceUrls: { type: string, maxLength: 2000 }
    BlockedUser:
      type: object
      additionalProperties: false
      required: [id, ownerId, blockedUserId, createdAt]
      properties:
        id: { type: string, format: uuid }
        ownerId: { type: string, format: uuid }
        blockedUserId: { type: string, format: uuid }
        reason: { type: string, maxLength: 500 }
        createdAt: { type: string, format: date-time }
    BlockUserInput:
      type: object
      additionalProperties: false
      required: [blockedUserId]
      properties:
        blockedUserId: { type: string, format: uuid }
        reason: { type: string, maxLength: 500 }
    TrustedContact:
      type: object
      additionalProperties: false
      required: [id, ownerId, name, createdAt]
      properties:
        id: { type: string, format: uuid }
        ownerId: { type: string, format: uuid }
        name: { type: string, maxLength: 120 }
        phone: { type: string, maxLength: 20 }
        email: { type: string, format: email, maxLength: 200 }
        relationship: { type: string, maxLength: 120 }
        createdAt: { type: string, format: date-time }
    CreateTrustedContactInput:
      type: object
      additionalProperties: false
      required: [name]
      properties:
        name: { type: string, maxLength: 120 }
        phone: { type: string, maxLength: 20 }
        email: { type: string, format: email, maxLength: 200 }
        relationship: { type: string, maxLength: 120 }
    PackageTracking:
      type: object
      additionalProperties: false
      required: [id, ownerId, trackingNumber, carrier, status, createdAt]
      properties:
        id: { type: string, format: uuid }
        ownerId: { type: string, format: uuid }
        trackingNumber: { type: string, maxLength: 80 }
        carrier: { type: string, maxLength: 80 }
        status: { type: string, enum: [PENDING, IN_TRANSIT, OUT_FOR_DELIVERY, DELIVERED, EXCEPTION] }
        estimatedDelivery: { type: string, format: date-time }
        lastUpdate: { type: string, format: date-time }
        createdAt: { type: string, format: date-time }
    CreatePackageTrackingInput:
      type: object
      additionalProperties: false
      required: [trackingNumber, carrier, status]
      properties:
        trackingNumber: { type: string, maxLength: 80 }
        carrier: { type: string, maxLength: 80 }
        status: { type: string, enum: [PENDING, IN_TRANSIT, OUT_FOR_DELIVERY, DELIVERED, EXCEPTION] }
        estimatedDelivery: { type: string, format: date-time }
    NearbyPlace:
      type: object
      additionalProperties: false
      required: [id, name, category, address, latitude, longitude, createdAt]
      properties:
        id: { type: string, format: uuid }
        name: { type: string, maxLength: 200 }
        category: { type: string, maxLength: 80 }
        address: { type: string, maxLength: 400 }
        latitude: { type: number, format: double }
        longitude: { type: number, format: double }
        rating: { type: number, format: double }
        distanceKm: { type: number, format: double }
        phone: { type: string, maxLength: 20 }
        createdAt: { type: string, format: date-time }
    EmergencyResource:
      type: object
      additionalProperties: false
      required: [id, name, category, address, description, latitude, longitude, createdAt]
      properties:
        id: { type: string, format: uuid }
        name: { type: string, maxLength: 200 }
        category: { type: string, maxLength: 80 }
        address: { type: string, maxLength: 400 }
        phone: { type: string, maxLength: 20 }
        hours: { type: string, maxLength: 120 }
        description: { type: string, maxLength: 4000 }
        latitude: { type: number, format: double }
        longitude: { type: number, format: double }
        createdAt: { type: string, format: date-time }
...
security:
  - bearerAuth: []